Implement first iteration of Upcoming Hikes component
